float for price

+++ database/migrations/2019_08_10_132304_create_lesson_exam_table.php
@@ -23,7 +23,7 @@
                 $table->string('exm')->nullable();
                 $table->string('title');
                 $table->string('description')->nullable();
-                $table->integer('price')->nullable()->default(0);
+                $table->float('price')->nullable()->default(0);
                 $table->string('answerSheet')->nullable();
                 $table->date('activeDate')->nullable();
                 $table->integer('duration')->nullable();
